+++ src/main.rs
@@ -1,23 +1,17 @@
 mod controller;
 mod drivers;
+mod handlers;
 mod types;
 
 use crate::drivers::{CloudWatchDriver, LokiDriver};
 use crate::types::LogDriver;
-use axum::{
-    body::{Body, Bytes},
-    extract::State,
-    http::{header::HeaderMap, Response, StatusCode},
-    response::IntoResponse,
-    routing::get,
-};
-use axum_prometheus::metrics::counter;
+use axum::routing::get;
 use axum_prometheus::PrometheusMetricLayerBuilder;
 use clap::Parser;
 use ring::hmac;
 use tokio::signal::{unix, unix::SignalKind};
 use tokio::sync::mpsc;
-use tracing::{debug, error, info, warn, Level};
+use tracing::{debug, info, Level};
 
 #[derive(Debug, Parser)]
 #[command(author, version, about, long_about = None)]
@@ -50,13 +44,6 @@ struct Args {
     loki_basic_auth_user: String,
     #[arg(long, env = "VERCEL_LOG_DRAIN_LOKI_PASS", default_value = "")]
     loki_basic_auth_pass: String,
-}
-
-#[derive(Debug, Clone)]
-struct AppState {
-    vercel_verify: String,
-    vercel_secret: hmac::Key,
-    log_queue: mpsc::UnboundedSender<types::Message>,
 }
 
 #[tokio::main]
@@ -94,7 +81,7 @@ async fn main() -> anyhow::Result<()> {
     tokio::spawn(async move {
         controller.run().await;
     });
-    let state = AppState {
+    let state = types::AppState {
         vercel_verify: args.vercel_verify,
         vercel_secret: hmac::Key::new(
             hmac::HMAC_SHA1_FOR_LEGACY_USE_ONLY,
@@ -106,11 +93,7 @@ async fn main() -> anyhow::Result<()> {
     let listen_address = format!("{}:{}", args.ip, args.port);
     let listener = tokio::net::TcpListener::bind(listen_address.clone()).await?;
 
-    let mut app = axum::Router::new()
-        .route("/", axum::routing::post(root))
-        .route("/health", axum::routing::get(health_check))
-        .route("/vercel", axum::routing::post(ingest))
-        .with_state(state);
+    let mut app = create_app(state);
 
     if args.enable_metrics {
         let (prometheus_layer, metric_handle) = PrometheusMetricLayerBuilder::new()
@@ -156,79 +139,10 @@ async fn shutdown_for_signals() {
     }
 }
 
-async fn root() -> impl IntoResponse {
-    Response::builder()
-        .status(StatusCode::OK)
-        .body(Body::empty())
-        .unwrap()
-}
-
-async fn ingest(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    body: Bytes,
-) -> impl IntoResponse {
-    debug!("received payload");
-
-    let signature = match headers.get("x-vercel-signature") {
-        Some(signature) => signature.to_str().unwrap(),
-        None => {
-            warn!("received payload without signature");
-            counter!("drain_recv_invalid_signature").increment(1);
-            return Response::builder()
-                .status(StatusCode::BAD_REQUEST)
-                .body(Body::empty())
-                .expect("Defined Responses to be infalliable.");
-        }
-    };
-    let body_string = match String::from_utf8(body.to_vec()) {
-        Ok(body_string) => body_string,
-        Err(e) => {
-            error!("received bad utf-8: {:?}", e);
-            counter!("drain_recv_bad_utf8").increment(1);
-            return Response::builder()
-                .status(StatusCode::NOT_ACCEPTABLE)
-                .body(Body::empty())
-                .expect("Defined Responses to be infalliable.");
-        }
-    };
-    let mut sig_bytes = [0u8; 20];
-    hex::decode_to_slice(signature, &mut sig_bytes).unwrap();
-    match hmac::verify(&state.vercel_secret, body_string.as_bytes(), &sig_bytes) {
-        Ok(_) => {}
-        Err(e) => {
-            error!("failed verifying signature: {:?}", e);
-            counter!("drain_failed_verify_signature").increment(1);
-            return Response::builder()
-                .status(StatusCode::UNPROCESSABLE_ENTITY)
-                .header("x-vercel-verify", state.vercel_verify)
-                .body(Body::empty())
-                .expect("Defined Responses to be infalliable.");
-        }
-    }
-    match serde_json::from_str::<types::VercelPayload>(&body_string) {
-        Ok(payload) => {
-            debug!("parsed payload, OK");
-            for message in payload.0 {
-                match state.log_queue.send(message) {
-                    Ok(_) => {}
-                    Err(e) => {
-                        error!("failed to queue log message to be sent to outputs: {:?}", e);
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            error!(payload = ?body_string, "failed parsing: {:?}", e);
-        }
-    }
-    return Response::builder()
-        .status(StatusCode::OK)
-        .header("x-vercel-verify", state.vercel_verify)
-        .body(Body::empty())
-        .expect("Defined Responses to be infalliable.");
-}
-
-async fn health_check() -> impl IntoResponse {
-    return StatusCode::OK;
+fn create_app(state: types::AppState) -> axum::Router {
+    return axum::Router::new()
+        .route("/", axum::routing::post(handlers::root))
+        .route("/health", axum::routing::get(handlers::health_check))
+        .route("/vercel", axum::routing::post(handlers::ingest))
+        .with_state(state);
 }
